Add support for hardwares that may not have the pcm format we want
This aims to fix https://github.com/Amjad50/dynwave/issues/2

Where the emulator is requesting PCM F32 format, but the device the user
have doesn't support that.

We can get around that by choosing a format the hardware support and
then converting each sample to this new format.

If the device has the same format, the `from_sample` should be optimized
away to nothing.

Signed-off-by: Amjad Alsharafi <26300843+Amjad50@users.noreply.github.com>